Fuctional CPU with interrupts and finished timer

+++ includes/instructions.h
@@ -7,14 +7,16 @@
 
 typedef struct {
     uint8_t mnemonic[16];
-    uint32_t length;
-    uint32_t cycles; // The lower bound if the num of cycles can vary (e.g. cond jump)
-    uint32_t (*func)(GB*);
+    uint8_t length;        // Length in bytes
+    uint8_t cycles;        // The lower bound if the num of cycles can vary (e.g. cond jump)
+    uint32_t (*func)(GB*);  // Function pointer to the instruction
 } Instruction;
 
 
+// Pushes a word into memory. Updates SP
 void push(GB *gb, uint16_t data);
 
+// Pops a word from memory. Updates SP
 uint16_t pop(GB *gb);
 
 #endif /* INSTRUCTIONS_H */
